Move all components to `lib/ui/layout`
