+++ src/lib/ui/layout/DetailRow.tsx
