+++ src/lib/ui/layout/DetailSection.tsx
