+++ src/lib/ui/layout/DetailSectionGroup.tsx
